UPDATE split design module into a subpackage, one file per algorithm
design.py (~935 lines) becomes a design/ package with screen,
greedy_substitution, beam_substitution, and greedy_marginalize each in
their own module, plus the shared _fast_tile_substitute numba kernel in
the private _substitute module. design/__init__.py re-exports all four so
`from tangermeme.design import X` is unchanged. tests/test_design.py is
split to mirror (test_screen, test_greedy_substitution,
test_beam_substitution, test_greedy_marginalize), and the design API page
documents each submodule. Pure move: no behavior change, full suite green.
