+++ tangermeme/design/__init__.py
@@ -1,0 +1,14 @@
+# design/__init__.py
+# Contact: Jacob Schreiber <jmschreiber91@gmail.com>
+
+from .screen import screen
+from .greedy_substitution import greedy_substitution
+from .beam_substitution import beam_substitution
+from .greedy_marginalize import greedy_marginalize
+
+__all__ = [
+	'screen',
+	'greedy_substitution',
+	'beam_substitution',
+	'greedy_marginalize',
+]
