--- tests/test_design.py
@@ -1,559 +1,0 @@
-# test_design.py
-# Contact: Jacob Schreiber <jmschreiber91@gmail.com>
-
-import torch
-import pytest
-
-from tangermeme.utils import random_one_hot
-from tangermeme.ersatz import substitute
-
-from tangermeme.predict import predict
-
-from tangermeme.design import screen
-from tangermeme.design import greedy_substitution
-from tangermeme.design import beam_substitution
-from tangermeme.design import greedy_marginalize
-
-from tangermeme.io import read_meme
-from tangermeme.utils import characters
-
-from .toy_models import SumModel
-from .toy_models import SmallDeepSEA
-
-from numpy.testing import assert_raises
-from numpy.testing import assert_array_almost_equal
-
-
-torch.manual_seed(0)
-torch.use_deterministic_algorithms(True, warn_only=True)
-
-
-@pytest.fixture
-def X():
-	return random_one_hot((1, 4, 100), random_state=0)
-
-
-@pytest.fixture
-def X_marg():
-	return random_one_hot((5, 4, 100), random_state=0)
-
-
-@pytest.fixture
-def motifs():
-	return [characters(m, force=True) for m in 
-		read_meme("tests/data/test.meme").values()]
-
-
-###
-
-
-def test_screen(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [10]
-
-	X_hat = screen(model, (4, 100), y, device=device, max_iter=1, random_state=0)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:, :, :20], [
-		[[1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0],
-         [0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1],
-         [0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0],
-         [0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0]]
-	])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.054611]])
-
-	##
-
-	X_hat = screen(model, (4, 100), y, device=device, max_iter=10, random_state=0)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:, :, :20], [
-		[[0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0],
-         [0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0],
-         [1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1]]
-	])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.03848]])
-
-
-def test_screen_summodel(device):
-	torch.manual_seed(0)
-	model = SumModel()
-	y = [10, 0, 0, 0]
-
-	X_hat = screen(model, (4, 10), y, device=device, max_iter=25, random_state=0)
-
-	assert X_hat.shape == (1, 4, 10)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:, :, :20], [
-		[[0, 1, 1, 1, 1, 1, 1, 0, 0, 1],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 1, 1, 0]]
-	])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[7., 0., 1., 2.]])
-
-
-def test_screen_n_best(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [10]
-	
-	X_hat = screen(model, (4, 100), y, device=device, n_best=7, 
-		max_iter=10, random_state=0)
-
-	assert X_hat.shape == (7, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:3, :, :10], [
-		[[0, 0, 0, 0, 0, 0, 1, 0, 0, 0],
-         [0, 1, 1, 1, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 1, 0, 1, 0, 1],
-         [1, 0, 0, 0, 1, 0, 0, 0, 1, 0]],
-
-        [[0, 0, 0, 1, 0, 0, 0, 0, 0, 0],
-         [1, 0, 0, 0, 0, 0, 0, 1, 0, 0],
-         [0, 1, 0, 0, 0, 0, 1, 0, 0, 0],
-         [0, 0, 1, 0, 1, 1, 0, 0, 1, 1]],
-
-        [[1, 0, 0, 0, 0, 0, 1, 0, 0, 1],
-         [0, 1, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 1, 0, 1, 0, 0],
-         [0, 0, 1, 1, 1, 0, 0, 0, 1, 0]]])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [
-		[-0.0385],
-        [-0.0439],
-        [-0.0454],
-        [-0.0457],
-        [-0.0466],
-        [-0.0496],
-        [-0.0510]
-	], 4)
-
-
-def test_screen_no_y(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-
-	X_hat = screen(model, (4, 100), device=device, random_state=0, max_iter=15)
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.085115]])
-
-
-def test_screen_custom_func(device):
-	torch.manual_seed(0)
-	def onlyAs(shape, random_state):
-		return random_one_hot(shape, probs=[1.0, 0.0, 0.0, 0.0])
-	
-	model = SmallDeepSEA()
-	y = [10]
-	
-	X_hat = screen(model, (4, 100), y, device=device, max_iter=10, 
-		random_state=0, func=onlyAs)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:, :, :20], [
-		[[1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
-	])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.0841]], 4)
-
-
-def test_screen_custom_kwargs(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [10]
-	
-	X_hat = screen(model, (4, 100), y, device=device, max_iter=1, 
-		additional_func_kwargs={'probs': [1, 0, 0, 0]})
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	assert_array_almost_equal(X_hat[:, :, :20], [
-		[[1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
-	])
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.0841]], 4)
-
-
-def test_screen_raises(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[0.5]]
-	
-	assert_raises(RuntimeError, screen, model, (5, 100), y, device=device)
-	assert_raises(ValueError, screen, model, (1, 4, 100), y, device=device)
-	assert_raises(ValueError, screen, model, (100,), y, device=device)
-
-
-###
-
-
-def test_greedy_substitution(X, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, motifs, device=device, max_iter=1)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.058557]])
-
-	##
-
-	X_hat = greedy_substitution(model, X, y, motifs, device=device, max_iter=4)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-	
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.018134]])
-
-
-def test_greedy_nucleotide_substitution(X, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, device=device, max_iter=1)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-	assert abs(X_hat - X).sum() == 2
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.076893]])
-
-	##
-
-	X_hat = greedy_substitution(model, X, y, device=device, max_iter=4)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-	assert abs(X_hat - X).sum() == 8
-	
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.057514]])
-
-
-def test_greedy_nucleotide_substitution_only_As(X, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	motifs = ['A']
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, motifs, device=device, max_iter=1, 
-		reverse_complement=False)
-	assert X_hat.sum(dim=(0, -1))[0] == X.sum(dim=(0, -1))[0] + 1
-
-	X_hat = greedy_substitution(model, X, y, motifs, device=device, max_iter=4, 
-		reverse_complement=False)
-	assert X_hat.sum(dim=(0, -1))[0] == X.sum(dim=(0, -1))[0] + 4
-
-
-def test_greedy_substitution_input_mask(X, device):
-	torch.manual_seed(0)
-	input_mask = torch.zeros(100, dtype=bool)
-	input_mask[54:58] = True
-
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, device=device, input_mask=input_mask,
-		max_iter=4)
-
-	assert_array_almost_equal(X[:, :, :54], X_hat[:, :, :54])
-	assert_array_almost_equal(X[:, :, 58:], X_hat[:, :, 58:])
-	assert abs(X[:, :, 54:58] - X_hat[:, :, 54:58]).sum() == 6
-
-
-def test_greedy_substitution_output_mask(X, device):
-	torch.manual_seed(0)
-	output_mask = torch.zeros(4, dtype=bool)
-	output_mask[3] = True
-
-	model = SumModel()
-	y = [[10, 10000, -26, 100]]
-
-	X_hat = greedy_substitution(model, X, y, device=device, output_mask=output_mask,
-		max_iter=101)
-
-	assert all(X_hat.sum(dim=(0, 2)) == torch.tensor([0, 0, 0, 100]))
-
-
-def test_greedy_substitution_reverse_complement(X, device):
-	torch.manual_seed(0)
-	output_mask = torch.zeros(4, dtype=bool)
-	output_mask[3] = True
-
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, ['A', 'C'], device=device, max_iter=101,
-		reverse_complement=False)
-
-	n_count = X.sum(dim=(0, 2))
-	n_hat_count = X_hat.sum(dim=(0, 2))
-	
-	assert n_count[0] < n_hat_count[0]
-	assert n_count[1] < n_hat_count[1]
-	assert n_count[2] > n_hat_count[2]
-	assert n_count[3] > n_hat_count[3]
-
-
-def test_greedy_substitution_no_y(X, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, device=device, max_iter=15)
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.01224]])
-
-
-###
-
-
-def test_beam_substitution(X, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = beam_substitution(model, X, y, motifs, beam_size=4, device=device,
-		max_iter=1)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=1).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.058557]])
-
-	##
-
-	X_hat = beam_substitution(model, X, y, motifs, beam_size=4, device=device,
-		max_iter=4)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-	y_hat = predict(model, X_hat, device=device)
-
-	assert_array_almost_equal(y_hat, [[-0.016707]])
-
-
-def test_beam_substitution_equals_greedy(X, motifs, device):
-	# A beam of width 1 is exactly the greedy substitution search.
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_greedy = greedy_substitution(model, X, y, motifs, device=device,
-		max_iter=4)
-	X_beam = beam_substitution(model, X, y, motifs, beam_size=1, device=device,
-		max_iter=4)
-
-	assert_array_almost_equal(X_beam, X_greedy)
-
-
-def test_beam_substitution_n_best(X, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = beam_substitution(model, X, y, motifs, beam_size=4, n_best=3,
-		device=device, max_iter=4)
-
-	assert X_hat.shape == (3, 4, 100)
-	assert X_hat.dtype == torch.int8
-
-	# Returned sequences are distinct and ranked from lowest to highest loss.
-	y_hat = predict(model, X_hat, device=device)
-	losses = ((torch.tensor([[10.0]]) - y_hat) ** 2).sum(dim=1)
-	assert (losses[:-1] <= losses[1:]).all()
-	assert len(torch.unique(X_hat, dim=0)) == 3
-
-
-def test_beam_substitution_max_iter_default(X, motifs, device):
-	# max_iter=-1 means no iteration limit; tol stops the search instead.
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = beam_substitution(model, X, y, motifs, beam_size=4, device=device)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert X_hat.dtype == torch.int8
-
-
-def test_beam_substitution_not_worse_than_greedy(X, motifs, device):
-	# The core promise: a wider beam carries the single best edit forward at
-	# every round, so its final loss can never exceed that of beam_size=1
-	# (which is greedy_substitution).
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = torch.tensor([[10.0]])
-
-	X_greedy = beam_substitution(model, X, y, motifs, beam_size=1, device=device,
-		max_iter=5)
-	X_beam = beam_substitution(model, X, y, motifs, beam_size=5, device=device,
-		max_iter=5)
-
-	loss_greedy = ((y - predict(model, X_greedy, device=device)) ** 2).sum()
-	loss_beam = ((y - predict(model, X_beam, device=device)) ** 2).sum()
-
-	assert loss_beam <= loss_greedy + 1e-3
-
-
-def test_beam_substitution_no_y(X, device):
-	# With y=None the objective is to maximize the model output, so the design
-	# must not decrease it relative to the starting sequence.
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-
-	y_before = predict(model, X, device=device)
-	X_hat = beam_substitution(model, X, beam_size=4, device=device, max_iter=4)
-	y_after = predict(model, X_hat, device=device)
-
-	assert X_hat.shape == (1, 4, 100)
-	assert y_after.sum() >= y_before.sum()
-
-
-def test_beam_substitution_input_mask(X, device):
-	# Only positions inside the mask may change (single-nucleotide edits).
-	torch.manual_seed(0)
-	input_mask = torch.zeros(100, dtype=bool)
-	input_mask[54:58] = True
-
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = beam_substitution(model, X, y, beam_size=4, device=device,
-		input_mask=input_mask, max_iter=4)
-
-	assert_array_almost_equal(X[:, :, :54], X_hat[:, :, :54])
-	assert_array_almost_equal(X[:, :, 58:], X_hat[:, :, 58:])
-
-
-def test_greedy_marginalize(X_marg, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10.0]]
-
-	X_hat = greedy_marginalize(model, X_marg, y, motifs, device=device, max_iter=1)
-
-	assert X_hat.shape == (4, 12)
-	assert X_hat.dtype == torch.int8
-	assert X_hat.sum(dim=0).min() == 1
-	assert all(torch.unique(X_hat) == torch.tensor([0, 1], dtype=torch.int8))
-
-
-###
-
-
-def test_screen_random_state_reproducibility(device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [10]
-
-	X_hat0 = screen(model, (4, 100), y, device=device, max_iter=1,
-		random_state=0)
-	X_hat1 = screen(model, (4, 100), y, device=device, max_iter=1,
-		random_state=0)
-
-	assert_array_almost_equal(X_hat0, X_hat1)
-
-
-def test_greedy_substitution_max_iter_default(X, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	X_hat = greedy_substitution(model, X, y, motifs, device=device)
-
-	assert_array_almost_equal(X_hat, X)
-
-
-def test_greedy_substitution_tol_no_improvement_stops(X, motifs, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-
-	y_current = predict(model, X, device=device)
-
-	X_hat = greedy_substitution(model, X, y_current, motifs, device=device,
-		max_iter=10, tol=1e-3)
-
-	assert_array_almost_equal(X_hat, X)
-
-
-def test_greedy_substitution_pre_encoded_motif_tensors(X, device):
-	torch.manual_seed(0)
-	model = SmallDeepSEA()
-	y = [[10]]
-
-	motif_tensor = torch.tensor(
-		[[1, 0], [0, 1], [0, 0], [0, 0]], dtype=torch.int8
-	)
-
-	assert_raises(Exception, greedy_substitution, model, X, y, [motif_tensor],
-		device=device, max_iter=1)
-
-
